Fixes to bounty and cargodepot event handling edge cases Made all reward fields decimal to allow for larger figures Handle edits to journal folders on the fly (still not right yet though...) Fixed missing handlers for some setting UI changes

+++ EDJournalQueue/Models/JournalEntryBounty.cs
@@ -8,7 +8,7 @@ namespace EDJournalQueue.Models
         #region Properties
 
         public string VictimFaction { get; set; }
-        public int TotalReward { get; set; }
+        public decimal TotalReward { get; set; }
         public string PilotName { get; set; }
 
         #endregion
